requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
argon2-cffi==20.1.0
async-timeout==3.0.1
attrs==20.1.0
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
crayons==0.3.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
discord.py==1.4.1
distlib==0.3.1
entrypoints==0.3
ffmpeg==1.4
ffmpeg-normalize==1.20.1
ffmpeg-python==0.2.0
ffprobe==0.5
filelock==3.0.12
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
install==1.3.3
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
mistune==0.8.4
multidict==4.7.6
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
notify2==0.3.1
oauthlib==3.1.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
piexif==1.1.3
Pillow==6.2.2
pipenv==2020.8.13
plotly==4.9.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
pycparser==2.20
pydub==0.24.1
Pygments==2.6.1
pymonocypher==3.1.0.0
PyNaCl==1.3.0
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.16.0
PySocks==1.7.1
pysodium==0.7.5
python-dateutil==2.8.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
stegano==0.9.8
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm==4.48.2
traitlets==5.0.4
tweepy==3.9.0
typing-extensions==3.7.4.3
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
yarl==1.5.1
youtube-dl==2020.7.28
zipp==3.1.0
